refactor out randomize utils, so that can more easily make a second randomized test
